requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
buildtools==1.0.6
causal-conv1d @ file:///root/autodl-tmp/Vim/causal_conv1d-1.1.3.post1%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=ea0743db0bf98b47ec53d890b5498d96f92896390efa0e57d233c84c885648f1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
conda-pack==0.8.1
constantly==23.10.4
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
dill==0.3.7
docker==6.1.3
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.2
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
grpcio==1.75.1
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.6
imageio==2.36.0
importlib-metadata==7.0.0
incremental==24.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
Mako==1.3.0
mamba-ssm @ file:///root/autodl-tmp/Vim/mamba_ssm-1.1.1%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=fb6a28285cc11a1bf7507b452903285a16ba36aec5f82f71e14223c83c4d484f
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.1.0
prettytable==3.9.0
protobuf==4.25.1
psutil==6.0.0
ptflops==0.7.3
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
redo==3.0.0
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.25.0
sentry-sdk==2.14.0
setproctitle==1.3.3
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.23
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.9.20
timm==0.4.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
torchaudio==2.1.1
torchprofile==0.0.4
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
trio==0.26.2
trio-websocket==0.11.1
triton==2.1.0
Twisted==24.7.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
wandb==0.21.0
wcwidth==0.2.12
websocket-client==1.8.0
Werkzeug==3.0.1
wsproto==1.2.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==7.0.3
